# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/login": {
            "post": {
                "tags": [
                    "Dashboard Api Auth User"
                ],
                "summary": "Login User in Dashboard",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "manager@app.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/dashboard/logout": {
            "post": {
                "tags": [
                    "Dashboard Api Auth User"
                ],
                "summary": "Auth Logout",
                "operationId": "d6f680d7d577dd7b278a44570b76d707",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/dashboard/categories": {
            "get": {
                "tags": [
                    "Dashboard Api Categories"
                ],
                "summary": "get all Category",
                "operationId": "91df637024d05c8e9272f2e3f329f0cb",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "filter search name categories",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dashboard Api Categories"
                ],
                "summary": "Add New Category",
                "operationId": "95a1bdb853b7d982e7f3f6ce945dc8b2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/dashboard/categories/{category_id}": {
            "put": {
                "tags": [
                    "Dashboard Api Categories"
                ],
                "summary": "Updated Category",
                "operationId": "66da321d6325e139f82cbe8c9a4d22b9",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dashboard Api Categories"
                ],
                "summary": "Delete Category",
                "operationId": "815787fb8dac8e80b1e5185c50f71972",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/dashboard/contacts": {
            "get": {
                "tags": [
                    "Dashboard Api Contacts"
                ],
                "summary": "get all Contacts",
                "operationId": "c7441b4d84640ab3ccd5837b47a06e14",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "filter search name or email or phone contacts",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/dashboard/contacts/{contact_id}": {
            "delete": {
                "tags": [
                    "Dashboard Api Contacts"
                ],
                "summary": "Delete Contact",
                "operationId": "d74931572318860392d74f4c28646220",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}